requirements.txt
autogen-agentchat===0.5.7
autogen-core
autogen-ext
python-dotenv
openai
gemini-api
tiktoken
pydantic
fastapi
uvicorn
websockets
jinja2
aiofiles